(F)unctional (F)eature (S)imple

This is the class that extracts simple features from functional columns. Note that it only operates on values that were actually observed and does not interpolate.

Format

R6Class object inheriting from PipeOpTaskPreprocSimple

Parameters

drop :: logical(1)
Whether to drop the original functional features and only keep the extracted features. Note that this does not remove the features from the backend, but only from the active column role feature. Initial value is TRUE.

affect_columns :: function | Selector | NULL
What columns the PipeOpTaskPreproc should operate on. See Selector for example functions. Defaults to NULL, which selects all features.

features :: list() | character()
A list of features to extract. Each element can be either a function or a string. If the element if is function it requires the following arguments: arg and value and returns a numeric. For string elements, the following predefined features are available: "mean", "max","min","slope","median","var". Initial is c("mean", "max", "min", "slope", "median", "var")

left :: numeric()
The left boundary of the window. Initial is -Inf. The window is specified such that the all values >=left and <=right are kept for the computations.

right :: numeric()
The right boundary of the window. Initial is Inf.

Naming

The new names generally append a _{feature} to the corresponding column name. However this can lead to name clashes with existing columns. This is solved as follows: If a column was called "x" and the feature is "mean", the corresponding new column will be called "x_mean". In case of duplicates, unique names are obtained using make.unique() and a warning is given.

Methods

Only methods inherited from PipeOpTaskPreprocSimple/ PipeOp

Super classes

mlr3pipelines::PipeOp -> mlr3pipelines::PipeOpTaskPreproc -> mlr3pipelines::PipeOpTaskPreprocSimple -> PipeOpFFS

Methods

Public methods

PipeOpFFS$new()

PipeOpFFS$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Initializes a new instance of this Class.

Usage

PipeOpFFS$new(id = "ffs", param_vals = list())

Arguments

id

(character(1))
Identifier of resulting object, default is "ffs".

param_vals

(named list)
List of hyperparameter settings, overwriting the hyperparameter settings that would

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpFFS$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3pipelines)
task = tsk("fuel")
po_fmean = po("ffs", features = "mean")
task_fmean = po_fmean$train(list(task))[[1L]]

# add more than one feature
pop = po("ffs", features = c("mean", "median", "var"))
task_features = pop$train(list(task))[[1L]]

# add a custom feature
po_custom = po("ffs", features = list(mean = function(arg, value) mean(value, na.rm = TRUE)))
task_custom = po_custom$train(list(task))[[1L]]