Flattens Functional Columns

Flattens Functional Columns

Flattens Functional Columns

Parameters

The parameters are the parameters inherited from PipeOpTaskPreprocSimple, as well as the following parameters:

as_features :: logical()
Whether to add the Flattened values to the features of the task.

selector :: function | Selector
Selector function, takes a Task as argument and returns a character of features to keep. The flattening is only applied to those columns.
See Selector for example functions. Default is selector_all(), which selects all of the functional` features.

Super classes

mlr3pipelines::PipeOp -> mlr3pipelines::PipeOpTaskPreproc -> mlr3pipelines::PipeOpTaskPreprocSimple -> PipeOpFlatFun

Methods

Public methods

PipeOpFlatFun$new()

PipeOpFlatFun$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Initializes a new instance of this Class.

Usage

PipeOpFlatFun$new(id = "flatfun", param_vals = list())

Arguments

id

()character(1))
Identifier of resulting object, default "ffe".

param_vals

(named list)
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpFlatFun$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.