Diffusion Tensor Imaging (DTI) Regression Task

This dataset contains two functional covariates and three scalar covariate. The goal is to predict the PASAT score. pasat represents the PASAT score at each vist. subject_id represents the subject ID. cca represents the fractional anisotropy tract profiles from the corpus callosum. sex indicates subject's sex. rcst represents the fractional anisotropy tract profiles from the right corticospinal tract. case indicates wether the subject has multiple sclerosis. Rows with NAs are removed.

This is a subset of the full dataset, which is contained in the package refund.

Format

R6::R6Class inheriting from TaskRegr.

Construction

mlr_tasks$get("dti")
tsk("dti")

References

Goldsmith, Jeff, Bobb, Jennifer, Crainiceanu, M C, Caffo, Brian, Reich, Daniel (2011). “Penalized functional regression.” Journal of Computational and Graphical Statistics, 20(4), 830--851.

Brain dataset courtesy of Gordon Kindlmann at the Scientific Computing and Imaging Institute, University of Utah, and Andrew Alexander, W. M. Keck Laboratory for Functional Brain Imaging and Behavior, University of Wisconsin-Madison.