Fuel Regression Task

This dataset contains two functional covariates and one scalar covariate. The goal is to predict the heat value of some fuel based on the ultraviolet radiation spectrum and infrared ray radiation and one scalar column called h2o.

This is a subset of the full dataset, which is contained in the package FDboost.

Format

R6::R6Class inheriting from TaskRegr.

Dictionary

This Task can be instantiated via the dictionary mlr_tasks or with the associated sugar function tsk():

mlr_tasks$get("fuel")
tsk("fuel")

Meta Information

Task type: “regr”

Dimensions: 129x4

Properties: -

Has Missings: FALSE

Target: “heatan”

Features: “NIR”, “UVVIS”, “h20”

References

Brockhaus, Sarah, Scheipl, Fabian, Hothorn, Torsten, Greven, Sonja (2015). “The functional linear array model.” Statistical Modelling, 15(3), 279–300.

See also