Fast Fourier transform features

This PipeOp extracts Fourier coefficients from functional columns. For more details, see stats::fft(), which is called internally. Only the one-sided spectrum is returned since the input is real-valued (Oppenheim and Schafer, 2010).

Parameters

The parameters are the parameters inherited from PipeOpTaskPreprocSimple, as well as the following parameters:

type :: character(1)
Which feature to extract from the Fourier coefficients. "amplitude" returns the magnitude. "phase" returns the phase shift in degrees (values in [-180, 180]). Initial value is "phase".

References

Oppenheim, V A, Schafer, W R (2010). Discrete-Time Signal Processing, 3rd edition. Pearson. ISBN 978-0131988422.

Super classes

mlr3pipelines::PipeOp -> mlr3pipelines::PipeOpTaskPreproc -> mlr3pipelines::PipeOpTaskPreprocSimple -> PipeOpFDAFourier

Methods

Public methods

PipeOpFDAFourier$new()

PipeOpFDAFourier$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Initializes a new instance of this Class.

Usage

PipeOpFDAFourier$new(id = "fda.fourier", param_vals = list())

Arguments

id

(character(1))
Identifier of resulting object, default is "fda.fourier".

param_vals

(named list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpFDAFourier$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = tsk("fuel")
po_fourier = po("fda.fourier")
task_fourier = po_fourier$train(list(task))[[1L]]
task_fourier$data()
#>       heatan    h2o NIR_fft_phase_1 NIR_fft_phase_2 NIR_fft_phase_3
#>        <num>  <num>           <num>           <num>           <num>
#>   1: 26.7810 2.3000               0       173.44750       179.85168
#>   2: 27.4720 3.0000               0       116.71370       104.35124
#>   3: 23.8400 2.0002               0      -115.67324       -92.81604
#>   4: 18.1680 1.8500               0       162.03218        72.80728
#>   5: 17.5170 2.3898               0       116.06575       102.32856
#>  ---                                                               
#> 125: 23.8340 2.1100               0       105.45411       107.83574
#> 126: 11.8050 1.6200             180        93.31335        91.50351
#> 127:  8.8315 1.4200             180        93.24224        91.57622
#> 128: 11.3450 1.4800               0       132.47256       125.13833
#> 129: 28.9940 2.5000               0      -112.11720       -96.22075
#>      NIR_fft_phase_4 NIR_fft_phase_5 NIR_fft_phase_6 NIR_fft_phase_7
#>                <num>           <num>           <num>           <num>
#>   1:       178.11849       177.79045      -178.94802       174.47297
#>   2:       101.15479        98.24095        96.48967        98.86313
#>   3:       -84.19807       -73.92503       -70.59741       -85.33020
#>   4:        70.36161        79.27097        70.11896        75.96298
#>   5:        95.37966        89.10806        84.72215        91.54888
#>  ---                                                                
#> 125:       101.96792       104.49473       110.73754       116.59146
#> 126:        91.11654        90.73911        89.78588        90.70623
#> 127:        89.92578        88.01634        83.89819        81.62355
#> 128:       111.81659       103.86099        92.64895        81.37822
#> 129:       -95.77575      -107.64834      -123.41718      -105.78833
#>      NIR_fft_phase_8 NIR_fft_phase_9 NIR_fft_phase_10 NIR_fft_phase_11
#>                <num>           <num>            <num>            <num>
#>   1:       172.73437      -153.61187       -152.95941       -165.10968
#>   2:       103.53553       102.99150        103.45602         97.36707
#>   3:       -85.73364       -65.44452       -100.24625       -115.28994
#>   4:        75.30619        70.86278         67.49782         53.17977
#>   5:        94.80790        91.46690         88.60239         86.82113
#>  ---                                                                  
#> 125:       114.23316       112.92366        118.46105        123.84853
#> 126:        90.46824        88.44631         86.38097         86.67259
#> 127:        84.24604        84.39934         85.11942         83.61495
#> 128:        73.44821        74.85036         82.04731         74.49771
#> 129:       -97.45380       -86.98545        -73.11540        -91.07522
#>      NIR_fft_phase_12 NIR_fft_phase_13 NIR_fft_phase_14 NIR_fft_phase_15
#>                 <num>            <num>            <num>            <num>
#>   1:       -165.03333       -175.01656       -169.99402        179.69982
#>   2:         97.44013         84.65771         81.96949         97.31739
#>   3:       -123.96085        169.82509         64.84589        -35.28738
#>   4:         56.49614         57.82378         58.50983         45.17721
#>   5:         92.32718         94.08597         92.15116         92.71050
#>  ---                                                                    
#> 125:        122.07680        118.83453        111.76608        105.43068
#> 126:         89.40290         88.07637         86.54696         89.19278
#> 127:         85.71009         87.29893         88.08282         86.22456
#> 128:         73.04738         62.41548         52.31288         61.16729
#> 129:       -104.24387       -129.76097       -144.07242       -158.29585
#>      NIR_fft_phase_16 NIR_fft_phase_17 NIR_fft_phase_18 NIR_fft_phase_19
#>                 <num>            <num>            <num>            <num>
#>   1:       178.600872        172.12942       -179.39202       -153.73587
#>   2:       114.521522         99.19947        119.87413        135.56018
#>   3:      -107.497989       -144.69542       -145.47113       -170.64634
#>   4:         9.598293        -78.60176        168.77008        135.29596
#>   5:        86.123173         80.54638         69.53938         77.81890
#>  ---                                                                    
#> 125:       123.854831        127.93432        129.75448        116.13075
#> 126:        91.299691         95.43466         96.64136         92.41033
#> 127:        85.564623         86.09575         90.05986         88.02184
#> 128:        52.020140         38.50868         42.46974         66.96264
#> 129:         7.080831        -60.50174        -60.37140        -85.11413
#>      NIR_fft_phase_20 NIR_fft_phase_21 NIR_fft_phase_22 NIR_fft_phase_23
#>                 <num>            <num>            <num>            <num>
#>   1:       -143.00813       -150.11777       -163.60535       -170.34202
#>   2:        135.84001        149.55110        177.37471       -177.20239
#>   3:        156.18199        163.78708        154.84105         95.18568
#>   4:        125.45304        117.91981        106.09320        122.76024
#>   5:         82.09379         87.80253         92.83351         84.93801
#>  ---                                                                    
#> 125:        104.49007         92.12109         90.84295        116.66956
#> 126:         90.50364         90.31255         87.73008         90.25271
#> 127:         90.92957         90.62199         91.57202         91.90733
#> 128:         75.98712         64.12148         69.69087         70.20678
#> 129:        -52.16048        -79.67477       -108.41878       -124.28681
#>      NIR_fft_phase_24 NIR_fft_phase_25 NIR_fft_phase_26 NIR_fft_phase_27
#>                 <num>            <num>            <num>            <num>
#>   1:       -166.79761       -165.39685       -167.04599       171.617602
#>   2:        155.72245        137.69566        175.58665      -166.299000
#>   3:         53.83769         48.44203         35.13092        -8.224618
#>   4:        102.73486         71.73796         57.39450       125.071036
#>   5:         79.48058        119.73741        135.45850       137.721546
#>  ---                                                                    
#> 125:        120.07694        106.55129         88.20245        65.182709
#> 126:         84.51337         84.11182         90.27655       105.062255
#> 127:         99.03068        101.25951        103.77801       102.259658
#> 128:         69.70345         63.23688         92.02002       100.565824
#> 129:        -85.72727        -75.31647        -99.19130       -35.015971
#>      NIR_fft_phase_28 NIR_fft_phase_29 NIR_fft_phase_30 NIR_fft_phase_31
#>                 <num>            <num>            <num>            <num>
#>   1:       161.224703       -167.55582       -138.23907      -148.281492
#>   2:      -179.244215        176.44324        166.21072       154.390647
#>   3:       119.020427         86.10644         30.98299        -2.817871
#>   4:       110.660682         93.33024         90.23872       100.672834
#>   5:       124.112835        110.13228        118.29310       125.276967
#>  ---                                                                    
#> 125:        43.530501         40.78546        121.87663       107.823079
#> 126:       115.970037        107.21170        104.91807       105.344287
#> 127:        98.297909         95.90867        101.52972       103.194924
#> 128:        84.230558         59.99833         50.61417        89.273846
#> 129:        -5.121416        -26.97642        -14.10566       -53.037478
#>      NIR_fft_phase_32 NIR_fft_phase_33 NIR_fft_phase_34 NIR_fft_phase_35
#>                 <num>            <num>            <num>            <num>
#>   1:       -160.52015       -158.61230       -163.69779       -178.81535
#>   2:        149.84091        135.63805        145.50568        153.84267
#>   3:        -34.35954        -79.67595       -144.68427        131.47458
#>   4:        132.10352        126.78008        108.95004        104.96829
#>   5:        132.71922        132.96308        129.28538        129.64922
#>  ---                                                                    
#> 125:        137.08259        -93.48902       -145.39134       -136.56230
#> 126:        111.33795        113.96656        115.14744        114.44951
#> 127:        106.91388        113.09847        113.03043        107.55403
#> 128:         87.64600         67.98845         61.06251         73.59174
#> 129:        -69.62615        -74.26423        -90.05111       -130.95856
#>      NIR_fft_phase_36 NIR_fft_phase_37 NIR_fft_phase_38 NIR_fft_phase_39
#>                 <num>            <num>            <num>            <num>
#>   1:       -177.35433       -179.49608        165.55715        131.97320
#>   2:        142.67250        130.13597        110.02292         99.10440
#>   3:         87.26384         41.64354        -32.61645        -39.39151
#>   4:         95.38813         77.64551         16.82048       -132.77401
#>   5:        124.85012        125.75315        122.33869        110.14322
#>  ---                                                                    
#> 125:       -158.29058       -175.28217        171.50901        153.15120
#> 126:        119.40983        124.49687        129.69374        128.81725
#> 127:        105.53121        109.86262        120.48368        120.31350
#> 128:         71.56051         69.83229         59.28978         70.23331
#> 129:        164.10175        136.22480        111.09429         86.53299
#>      NIR_fft_phase_40 NIR_fft_phase_41 NIR_fft_phase_42 NIR_fft_phase_43
#>                 <num>            <num>            <num>            <num>
#>   1:        104.93592         34.12455       -56.367718       152.874825
#>   2:         84.99438         62.27805        38.188520        -2.540622
#>   3:        148.03123         73.75881        -8.393329       -63.819646
#>   4:       -154.34886       -167.85734       179.550254       163.031505
#>   5:        113.77510        120.95422       130.993951       145.835683
#>  ---                                                                    
#> 125:        146.14777        139.45453       121.531796       121.523624
#> 126:        124.15402        124.90206       119.778357       114.544393
#> 127:        118.49648        118.30808       121.458898       119.199999
#> 128:         88.63098         85.40419       134.388993       140.734451
#> 129:         52.96510         21.49278        -9.652622       -19.256845
#>      NIR_fft_phase_44 NIR_fft_phase_45 NIR_fft_phase_46 NIR_fft_phase_47
#>                 <num>            <num>            <num>            <num>
#>   1:         49.16091        -14.64593        -62.08125        -87.59685
#>   2:        -26.48261        -12.43612        -58.20153       -116.37668
#>   3:        -97.81415       -134.67020       -151.95734        152.31029
#>   4:        157.03010        148.58955        156.38365        142.87814
#>   5:        150.40733        146.85186        151.82633        140.25443
#>  ---                                                                    
#> 125:        163.10306        155.56810        151.45733        129.05613
#> 126:        116.93926        114.86139        126.54033        131.64658
#> 127:        124.01507        129.48911        129.80706        139.77969
#> 128:        140.91454        140.85661        140.04724        112.97721
#> 129:        -43.72056        -68.42115        -70.65943       -112.13157
#>      NIR_fft_phase_48 NIR_fft_phase_49 NIR_fft_phase_50 NIR_fft_phase_51
#>                 <num>            <num>            <num>            <num>
#>   1:        -74.03476        -70.02324        -91.74418       -117.35801
#>   2:       -153.45919       -148.15530        156.36466         89.38996
#>   3:       -147.78544       -150.17433        170.82451        160.38134
#>   4:        152.26184        164.15744        172.91940        164.08309
#>   5:        149.17964        164.74776        153.07013        149.81827
#>  ---                                                                    
#> 125:        109.21578        101.06143         73.36174         43.06520
#> 126:        129.27984        115.69573        113.42184        131.09910
#> 127:        140.97653        142.20945        143.08930        137.66912
#> 128:        124.93690        141.07341        150.36388        140.22571
#> 129:       -132.66148        -36.71525        -59.06393        -64.44844
#>      NIR_fft_phase_52 NIR_fft_phase_53 NIR_fft_phase_54 NIR_fft_phase_55
#>                 <num>            <num>            <num>            <num>
#>   1:      -155.524218        125.89102         45.60628         27.63819
#>   2:      -152.507167        159.98450        125.19001       -167.94577
#>   3:       136.131218       -171.97641       -169.13559        162.03533
#>   4:       147.500917        136.72649        150.04078        148.55418
#>   5:       146.524569        136.89929        140.18130        145.11432
#>  ---                                                                    
#> 125:         4.638984        -37.24409        -74.12306        -93.79020
#> 126:       160.899288        163.97299        146.08649        141.94460
#> 127:       134.662913        131.20172        132.04472        139.93516
#> 128:       138.460694        146.33951        141.01724        121.60325
#> 129:       -86.101012       -124.91041       -156.46214        158.16260
#>      NIR_fft_phase_56 NIR_fft_phase_57 NIR_fft_phase_58 NIR_fft_phase_59
#>                 <num>            <num>            <num>            <num>
#>   1:        -2.709553        -30.22736        -63.77113        -48.05306
#>   2:      -177.912506        136.72945         95.21097         66.64860
#>   3:       153.334883        144.75787        144.72264        168.86875
#>   4:       141.735535        169.58408        165.36319        152.54513
#>   5:       151.229999        162.57567        149.79278        149.78789
#>  ---                                                                    
#> 125:      -127.993202        -47.47776        -59.70475        -91.31505
#> 126:       153.243937        148.22248        143.04229        136.15133
#> 127:       137.586540        144.68001        144.94887        149.39400
#> 128:       111.135959        118.69345        123.91983        122.96391
#> 129:       127.142564        104.16209         79.08688         57.05493
#>      NIR_fft_phase_60 NIR_fft_phase_61 NIR_fft_phase_62 NIR_fft_phase_63
#>                 <num>            <num>            <num>            <num>
#>   1:        -59.29818        -68.49615        -82.68053       -86.160540
#>   2:         27.18379        -10.82613        -66.52730      -116.575738
#>   3:        171.70865        149.23753        136.57691       113.369821
#>   4:        151.84668        149.85790        146.96022       137.401685
#>   5:        166.50067        174.32485        172.36971       179.975300
#>  ---                                                                    
#> 125:       -105.21662       -120.44379       -135.21518      -143.707724
#> 126:        152.45006        146.91594        140.64117       148.385504
#> 127:        143.67174        138.05361        144.93112       144.685252
#> 128:        132.66308        135.11727        145.40312       133.435900
#> 129:         39.21856         55.52137         32.59830        -8.406124
#>      NIR_fft_phase_64 NIR_fft_phase_65 NIR_fft_phase_66 NIR_fft_phase_67
#>                 <num>            <num>            <num>            <num>
#>   1:        -85.42985        -15.13083        -42.16833        -71.32274
#>   2:        156.27155         94.78116         48.00431         23.95389
#>   3:         82.65556         72.33356         82.73996         63.88113
#>   4:        150.85814        157.62109        144.72525        130.38813
#>   5:       -179.74184       -176.13160       -171.37209        172.82871
#>  ---                                                                    
#> 125:       -139.57043       -157.20217        -82.58340        -90.60403
#> 126:        147.83412        148.71496        145.39182        119.21206
#> 127:        140.69046        136.42281        128.23822        122.00367
#> 128:        133.91909        141.94672        168.74788        162.96968
#> 129:        -47.29434        -55.95092        -68.04004        -78.51401
#>      NIR_fft_phase_68 NIR_fft_phase_69 NIR_fft_phase_70 NIR_fft_phase_71
#>                 <num>            <num>            <num>            <num>
#>   1:       -72.141059        -87.42786       -110.19552       -142.78216
#>   2:         4.997223         26.74644        -16.95906        -74.59470
#>   3:        74.323746         42.74316         34.47777         26.57199
#>   4:       101.307229        130.58012        124.32900        110.66451
#>   5:       149.525644        130.20530        114.05989        164.87662
#>  ---                                                                    
#> 125:      -108.899311       -101.60663        -69.41342        -63.24923
#> 126:       111.643992        123.91670        134.88248        179.15726
#> 127:       149.360998        143.21360        140.04293        134.10662
#> 128:       150.587035        143.21265        136.19548        121.73734
#> 129:       -87.082325       -121.52703       -150.28363       -162.61954
#>      NIR_fft_phase_72 NIR_fft_phase_73 NIR_fft_phase_74 NIR_fft_phase_75
#>                 <num>            <num>            <num>            <num>
#>   1:      -139.351610       -141.51605       -161.91949        129.81639
#>   2:      -114.535966       -154.90336       -168.07673        173.57457
#>   3:         3.072884        -48.12782        -70.39274        -69.47274
#>   4:       161.842060        128.65678         80.61053        -96.31275
#>   5:       176.975360        174.18180        161.02736        159.75528
#>  ---                                                                    
#> 125:       -81.382531       -126.07733       -136.67254       -158.97628
#> 126:       163.771176        179.25843        162.13031        176.71856
#> 127:       138.594266        130.39123        124.58488        111.15476
#> 128:       104.710595         88.46872        166.03628        157.43020
#> 129:      -166.421742        175.19952        146.09405        115.36481
#>      NIR_fft_phase_76 NIR_fft_phase_77 NIR_fft_phase_78 NIR_fft_phase_79
#>                 <num>            <num>            <num>            <num>
#>   1:        -72.45956        -65.69843      -103.478892       -117.54425
#>   2:        118.75430         61.62674        -2.698053        -44.79555
#>   3:        -86.45798        -46.60141       -35.956852        -71.88386
#>   4:       -162.66807       -171.47849       133.398824        112.21394
#>   5:        174.31302        177.90410       170.661543        166.58043
#>  ---                                                                    
#> 125:        135.70773         86.77838         9.145381        -26.09963
#> 126:        163.49552        148.82119      -177.557794       -179.51106
#> 127:         98.08330        151.66561       150.693790        132.14938
#> 128:        156.23327        160.08219       159.197731        162.78889
#> 129:         97.07651         91.82229       106.947884        165.44845
#>      NIR_fft_phase_80 NIR_fft_phase_81 NIR_fft_phase_82 NIR_fft_phase_83
#>                 <num>            <num>            <num>            <num>
#>   1:       -145.34267        179.21764        147.62979        106.32195
#>   2:        -59.06562       -100.01251        -50.38030        -63.14294
#>   3:        -95.99931       -101.45290        -97.48428        -64.53410
#>   4:        -72.27407       -128.68458       -111.95816       -126.11649
#>   5:        169.83275        169.07629        150.56706        133.36016
#>  ---                                                                    
#> 125:        -75.12321       -138.60780       -152.76286        169.17545
#> 126:       -175.93818        164.35637        142.45913        131.64567
#> 127:        142.10201        151.69353        148.14140        129.27223
#> 128:        170.81021        159.20199        162.33134       -173.15806
#> 129:        128.11761         71.32056          4.75263        -66.57901
#>      NIR_fft_phase_84 NIR_fft_phase_85 NIR_fft_phase_86 NIR_fft_phase_87
#>                 <num>            <num>            <num>            <num>
#>   1:         64.11112         15.26643       -147.09807        148.20526
#>   2:        -45.07688        -84.69475       -108.24882       -126.09409
#>   3:       -108.36419       -137.94481       -137.39419        167.79121
#>   4:       -147.90190        177.10371       -162.95546        178.28282
#>   5:        127.31915        139.25411        169.94568       -176.10686
#>  ---                                                                    
#> 125:        119.31300         71.27420         36.13458        -60.41441
#> 126:        148.89819        161.68470        161.22542        144.34109
#> 127:        110.36461        147.85897        146.33853        135.65254
#> 128:       -160.52790       -158.53074       -174.30518        164.26966
#> 129:       -119.28101       -162.18620        170.79562        146.09110
#>      NIR_fft_phase_88 NIR_fft_phase_89 NIR_fft_phase_90 NIR_fft_phase_91
#>                 <num>            <num>            <num>            <num>
#>   1:         127.3213         82.43245        37.026642         -2.25383
#>   2:        -100.3866       -121.43023      -119.163404       -123.54750
#>   3:          58.2013         14.30961       -21.396725        -48.63666
#>   4:         167.7983        169.57559       155.694441        125.41465
#>   5:         170.5740        150.47603       148.229296        154.80395
#>  ---                                                                    
#> 125:        -157.7679        169.02617       135.515389        114.03420
#> 126:         154.7873        161.11577       160.312149        157.65385
#> 127:         139.5803        158.00434       161.773896        163.74418
#> 128:         124.9864        111.88803      -139.873999       -173.22844
#> 129:         117.3256         62.48112         5.569581        -46.19286
#>      NIR_fft_phase_92 NIR_fft_phase_93 NIR_fft_phase_94 NIR_fft_phase_95
#>                 <num>            <num>            <num>            <num>
#>   1:        -34.30044       -127.68094        177.05989       144.127556
#>   2:       -129.55777       -148.39407       -175.79654      -141.212541
#>   3:        -76.34779        -79.33004        -96.21253      -122.051833
#>   4:         96.77872         78.49653         62.63084        62.419286
#>   5:        149.13175        150.01451        138.73011       126.164992
#>  ---                                                                    
#> 125:        102.20417         73.12738         37.15184        -9.148259
#> 126:        139.57331        129.10493        149.15978       158.064713
#> 127:        147.55242        148.96918        148.45630       150.981193
#> 128:        172.90582        152.94054        117.42920        76.030190
#> 129:        -94.55393       -118.26849       -154.81850       159.010049
#>      NIR_fft_phase_96 NIR_fft_phase_97 NIR_fft_phase_98 NIR_fft_phase_99
#>                 <num>            <num>            <num>            <num>
#>   1:        106.46183         71.22881        12.583710      -55.8401125
#>   2:       -142.23584       -144.69795      -147.531596     -157.4147425
#>   3:       -124.49197       -111.17324      -121.023926     -112.1763184
#>   4:         75.50106         38.06368         7.389332      -32.3080573
#>   5:        128.45914        120.68446       116.505546      168.6501370
#>  ---                                                                    
#> 125:       -176.75907        115.75352        64.066943       -0.7778902
#> 126:        151.84302        147.86692       141.620956      136.6516777
#> 127:        149.48538        141.77206       144.006618      138.0207535
#> 128:        -13.06371        -69.18011      -103.125776      -83.0244179
#> 129:        118.81984        133.86430       114.501109       66.1567632
#>      NIR_fft_phase_100 NIR_fft_phase_101 NIR_fft_phase_102 NIR_fft_phase_103
#>                  <num>             <num>             <num>             <num>
#>   1:         22.892756         -29.07838          114.7900          64.74264
#>   2:       -157.072714        -160.92439         -170.1343        -166.54288
#>   3:       -135.024641        -100.21491         -106.7938         -90.92509
#>   4:        -99.049165        -146.11485         -176.1553         137.14289
#>   5:        165.717258         179.64685          162.8368         133.07122
#>  ---                                                                        
#> 125:        113.949153        -159.13512          176.1049         141.50097
#> 126:        148.110081         149.33958          148.6637         179.86486
#> 127:        134.654848         122.35864          162.2617         165.57220
#> 128:        -62.351718         -98.51269         -123.2763        -146.20532
#> 129:          9.752682         151.95254          109.1730          69.75263
#>      NIR_fft_phase_104 NIR_fft_phase_105 NIR_fft_phase_106 NIR_fft_phase_107
#>                  <num>             <num>             <num>             <num>
#>   1:          44.68910        -25.370330        -175.45201        100.251396
#>   2:        -177.84738       -176.071612         158.09658        130.367423
#>   3:         -89.62468        -98.997904        -109.66966       -117.758827
#>   4:         109.57454         84.880811          55.54125         20.654349
#>   5:         113.17818        107.657579         -20.07143       -114.909413
#>  ---                                                                        
#> 125:         174.01716        110.775446          85.35599         -4.396741
#> 126:         173.81690       -142.676959        -157.77528       -165.672282
#> 127:        -167.43248        177.953233         176.89785        172.150915
#> 128:        -125.99310       -121.620629        -140.96362       -168.697414
#> 129:          32.27561          7.434013         -35.41220        -68.631554
#>      NIR_fft_phase_108 NIR_fft_phase_109 NIR_fft_phase_110 NIR_fft_phase_111
#>                  <num>             <num>             <num>             <num>
#>   1:          36.57908         -18.03107        -47.083809         -99.82526
#>   2:         168.98295         151.06632        105.792619          89.62970
#>   3:        -130.23429        -154.05388       -154.506499        -143.80231
#>   4:         -17.83654         -11.07403         -7.971161         -17.76031
#>   5:        -142.84379        -145.83901       -172.683881         163.64763
#>  ---                                                                        
#> 125:        -114.94126        -146.70284       -133.310798        -157.50778
#> 126:        -177.58023        -179.42995       -164.851538        -160.46363
#> 127:         168.53127         164.96058        171.033512         168.03380
#> 128:         157.46026         148.01881       -162.824880        -174.32830
#> 129:        -122.26192         170.45468        107.800217          59.38545
#>      NIR_fft_phase_112 NIR_fft_phase_113 NIR_fft_phase_114 NIR_fft_phase_115
#>                  <num>             <num>             <num>             <num>
#>   1:        -125.87149        -157.76225          25.76726        -59.921838
#>   2:         129.21736          96.05284        -104.18771       -149.407898
#>   3:        -158.42323        -170.44326        -174.11515       -168.737248
#>   4:         -42.68907         -60.24405        -155.08363         71.667154
#>   5:        -178.94482        -159.50524        -166.09155        163.758595
#>  ---                                                                        
#> 125:        -164.77047        -169.38685        -156.76092       -161.026146
#> 126:        -162.93942        -166.92937        -169.90194       -177.049879
#> 127:         171.84868        -171.22714        -159.53653       -173.237861
#> 128:         145.73639         145.66090         112.87183        -41.859135
#> 129:          25.47647          22.71851          23.33658          2.053627
#>      NIR_fft_phase_116 UVVIS_fft_phase_1 UVVIS_fft_phase_2 UVVIS_fft_phase_3
#>                  <num>             <num>             <num>             <num>
#>   1:        -148.57264                 0          67.18685          98.57592
#>   2:         174.61184               180          74.89705          82.21205
#>   3:        -173.44842               180          88.88101          73.75813
#>   4:          18.22147               180          56.67719          78.93985
#>   5:         135.34535               180          68.12330          83.19343
#>  ---                                                                        
#> 125:        -165.41473               180          54.92245          98.93474
#> 126:        -176.75550               180          17.26104         -23.83343
#> 127:        -177.44111               180          10.58915         -28.20270
#> 128:        -139.77291                 0          17.39356         171.05680
#> 129:         -10.77290               180          91.15874          87.04682
#>      UVVIS_fft_phase_4 UVVIS_fft_phase_5 UVVIS_fft_phase_6 UVVIS_fft_phase_7
#>                  <num>             <num>             <num>             <num>
#>   1:         92.868964         96.665145          94.11634         106.18532
#>   2:         86.138848        101.608183         121.05059         127.28131
#>   3:         13.626115        -66.041592         152.58885         123.46584
#>   4:        107.358030        126.150308         166.24105         159.67414
#>   5:         55.639290         49.645747          80.44402          84.06376
#>  ---                                                                        
#> 125:        118.232340        162.676127         175.66908        -150.26027
#> 126:        -92.739004        119.796753         124.41039         157.71703
#> 127:         -4.627931         -9.788832          28.92564          28.30431
#> 128:        116.175769        109.627727          83.59546          48.40181
#> 129:         95.738038        123.087891         122.36515         107.46145
#>      UVVIS_fft_phase_8 UVVIS_fft_phase_9 UVVIS_fft_phase_10 UVVIS_fft_phase_11
#>                  <num>             <num>              <num>              <num>
#>   1:         103.49437         95.369553          94.400731          120.66875
#>   2:         108.82819         92.094931         133.095047          162.42197
#>   3:          73.19248         16.626550          -1.508445          -24.48229
#>   4:          55.07123         -8.235251         -53.372231          -61.84871
#>   5:          94.61765         78.529868          92.457937           94.91182
#>  ---                                                                          
#> 125:        -177.71977         61.717436          55.006020           56.02124
#> 126:        -161.52533       -136.903334        -151.254321          169.90153
#> 127:         -16.31684       -108.152447        -147.490693          140.81845
#> 128:        -104.15276       -160.985692         176.093513          102.17941
#> 129:         101.52391        172.654162         168.462339          148.00614
#>      UVVIS_fft_phase_12 UVVIS_fft_phase_13 UVVIS_fft_phase_14
#>                   <num>              <num>              <num>
#>   1:          116.59966           99.10074           84.34227
#>   2:          157.13172          119.34531          103.51327
#>   3:          -45.29235          -60.84420         -110.14573
#>   4:          -40.89234           37.05523         -165.36909
#>   5:          119.80609          121.19173          103.88889
#>  ---                                                         
#> 125:           49.51210           28.31010          -13.72666
#> 126:           68.58638           41.07073          109.91022
#> 127:          129.21103          -80.28419          -44.92706
#> 128:           28.06081          -55.78244         -105.38602
#> 129:          130.00695          142.11983          122.11455
#>      UVVIS_fft_phase_15 UVVIS_fft_phase_16 UVVIS_fft_phase_17
#>                   <num>              <num>              <num>
#>   1:           77.74343          29.347839           50.74983
#>   2:          130.83441         140.826578          111.35242
#>   3:           84.29893         -70.191770         -162.65393
#>   4:          155.06330         131.562192          162.02583
#>   5:           85.19935          86.868956         -173.63595
#>  ---                                                         
#> 125:          151.80818         134.845531          -34.76286
#> 126:           73.90894        -142.451201          137.11258
#> 127:            9.28168          -4.302251         -162.36725
#> 128:          178.35666         123.407567          123.29349
#> 129:           99.69218          76.474599          145.18142
#>      UVVIS_fft_phase_18 UVVIS_fft_phase_19 UVVIS_fft_phase_20
#>                   <num>              <num>              <num>
#>   1:           62.66331           52.01568          62.983329
#>   2:         -175.62058         -151.54747         173.839276
#>   3:          169.77967          148.54976          -7.559888
#>   4:          163.17264          156.93321         170.311642
#>   5:          140.70588          126.20138         106.526554
#>  ---                                                         
#> 125:         -145.64719         -177.74103         153.698337
#> 126:          113.09076          113.44600         104.628950
#> 127:          155.66284           88.30543         143.136314
#> 128:          119.15165          128.68079         135.405083
#> 129:          152.31223          178.01567         155.485157
#>      UVVIS_fft_phase_21 UVVIS_fft_phase_22 UVVIS_fft_phase_23
#>                   <num>              <num>              <num>
#>   1:          104.45830          128.01098           77.50643
#>   2:          158.59561         -173.18118         -171.91199
#>   3:          -44.82791         -133.85983          177.03021
#>   4:         -173.32104         -142.04719          -55.38045
#>   5:           94.42751           88.89019          139.61034
#>  ---                                                         
#> 125:          111.91305           95.32639          -28.88643
#> 126:           82.25834           41.92130           35.16016
#> 127:           88.81697          -17.78375          -56.20207
#> 128:           99.83074          131.51752         -178.25746
#> 129:          127.91617          136.76068          158.79517
#>      UVVIS_fft_phase_24 UVVIS_fft_phase_25 UVVIS_fft_phase_26
#>                   <num>              <num>              <num>
#>   1:           90.05458        138.3279744          114.56336
#>   2:         -177.97443        178.5637189          137.22897
#>   3:          156.98746        150.7374450          169.80191
#>   4:          -27.24728          0.5989558           51.04297
#>   5:          152.82978        165.7930500          133.07168
#>  ---                                                         
#> 125:           20.01054         45.1062158           28.44744
#> 126:           32.21570         33.6145609           20.80944
#> 127:         -115.68521        116.7076995           98.82723
#> 128:          124.58955         87.4719114           25.44782
#> 129:          152.85752        129.6018830          119.68303
#>      UVVIS_fft_phase_27 UVVIS_fft_phase_28 UVVIS_fft_phase_29
#>                   <num>              <num>              <num>
#>   1:         100.824175          69.944150           96.92901
#>   2:          79.986580          35.627906         -154.02653
#>   3:         -75.063720         -95.086791         -144.67991
#>   4:          -2.895990         -82.918441         -161.25630
#>   5:          70.962822          -7.975742         -135.18332
#>  ---                                                         
#> 125:           2.007382         -62.064766         -146.83183
#> 126:          76.137804          96.693532           33.11847
#> 127:         -10.732593        -139.797193          -78.08293
#> 128:         -40.161692        -126.012494          137.72569
#> 129:          89.695687          61.524367          -87.90812
#>      UVVIS_fft_phase_30 UVVIS_fft_phase_31 UVVIS_fft_phase_32
#>                   <num>              <num>              <num>
#>   1:         119.123988          155.50394        -160.621125
#>   2:         166.499614          106.08148        -140.578331
#>   3:          88.759514           28.26133         -34.286348
#>   4:         106.896697           77.40805           3.405828
#>   5:         127.690326           81.13769          31.218399
#>  ---                                                         
#> 125:         167.332190          169.29532         169.288630
#> 126:          -4.409097          -34.45515         -46.041419
#> 127:         -19.261556          -52.69004        -104.826686
#> 128:         128.298505           11.04897        -133.068090
#> 129:        -133.263252         -175.79438         -89.314338
#>      UVVIS_fft_phase_33 UVVIS_fft_phase_34 UVVIS_fft_phase_35
#>                   <num>              <num>              <num>
#>   1:        -145.391757         -150.07913         -157.37557
#>   2:         168.703567         -178.15676          -67.72572
#>   3:         -81.046491         -103.09156         -161.43934
#>   4:         -61.520822         -139.54999         -153.79379
#>   5:           1.767025          140.79332          126.26650
#>  ---                                                         
#> 125:         178.508657          175.94199         -177.65483
#> 126:         177.899333          117.95966           64.27815
#> 127:        -174.326907          154.01780          141.52247
#> 128:        -104.017423          -91.19224          -83.17242
#> 129:         -43.535837          -45.92784          -88.39299
#>      UVVIS_fft_phase_36 UVVIS_fft_phase_37 UVVIS_fft_phase_38
#>                   <num>              <num>              <num>
#>   1:          149.61965        128.2555516          89.589884
#>   2:          -90.21433         21.9908474          39.437492
#>   3:           78.06461        -17.5446824           3.268597
#>   4:          178.73052        115.6815531          75.596921
#>   5:          141.76681        147.6693984          92.393486
#>  ---                                                         
#> 125:         -177.90681        177.1238284         165.647955
#> 126:           52.80869         90.3421911         115.954482
#> 127:          -99.07819       -158.3685960          69.427025
#> 128:          -77.63343       -144.5329689         169.004606
#> 129:           12.65741         -0.7005146          12.152297
#>      UVVIS_fft_phase_39 UVVIS_fft_phase_40 UVVIS_fft_phase_41
#>                   <num>              <num>              <num>
#>   1:         121.688010         143.720384          131.32837
#>   2:           1.008589         -65.168877         -135.58480
#>   3:          23.954185           5.487295            6.73490
#>   4:         133.979310         111.999914          147.20736
#>   5:          59.896225          53.266684           78.35370
#>  ---                                                         
#> 125:         131.275392          84.600489           43.12281
#> 126:          95.566482         104.371921           98.48346
#> 127:          -1.256891         -36.549020          -66.27006
#> 128:         153.018487        -127.206134         -171.57569
#> 129:          70.841780          74.054928           38.75901
#>      UVVIS_fft_phase_42 UVVIS_fft_phase_43 UVVIS_fft_phase_44
#>                   <num>              <num>              <num>
#>   1:         123.601705          95.823284           71.83190
#>   2:        -137.596238        -170.109947          116.45862
#>   3:          -4.822462         131.406663          107.84979
#>   4:         160.600358         140.385018          115.86361
#>   5:          89.061415          94.055223           73.06459
#>  ---                                                         
#> 125:           1.190095          -3.498167           92.02964
#> 126:          83.546272          69.045713           36.17057
#> 127:         176.904263         150.840496          155.15892
#> 128:          98.496732          30.040545         -136.24387
#> 129:          12.358293         -69.277129          -35.61031
#>      UVVIS_fft_phase_45 UVVIS_fft_phase_46 UVVIS_fft_phase_47
#>                   <num>              <num>              <num>
#>   1:           20.58410         30.3458029         -170.89896
#>   2:           43.45091         -0.6330901          -89.66718
#>   3:           29.40285       -127.6689735         -138.20889
#>   4:           65.05245         90.3272904          169.79302
#>   5:           17.98874        -91.4531487         -151.66172
#>  ---                                                         
#> 125:           98.87264         87.7939194           78.84483
#> 126:          -47.78449       -102.9347642         -159.16079
#> 127:          133.42486         -6.3540718          -53.93803
#> 128:         -156.71141       -174.7453442          165.57567
#> 129:          -43.01238        -65.0004324          -96.27935
#>      UVVIS_fft_phase_48 UVVIS_fft_phase_49 UVVIS_fft_phase_50
#>                   <num>              <num>              <num>
#>   1:        -175.407089          147.42820          143.99022
#>   2:        -160.064472          136.23022           40.40799
#>   3:        -130.602402         -125.11680          106.94729
#>   4:         149.888103          156.17802          149.71578
#>   5:         166.804203          130.00239           94.94662
#>  ---                                                         
#> 125:           9.627159          -56.55134          -91.76888
#> 126:        -175.718455         -132.57568         -158.38892
#> 127:         -70.708588          -49.31566          -39.26415
#> 128:         147.544572           94.72714           42.49045
#> 129:        -102.527654          -92.33079          -52.79842
#>      UVVIS_fft_phase_51 UVVIS_fft_phase_52 UVVIS_fft_phase_53
#>                   <num>              <num>              <num>
#>   1:          120.99353         -129.18523         -99.887240
#>   2:          -17.80893           88.74101          57.848051
#>   3:           43.77982           43.95687           6.432881
#>   4:         -160.21211         -172.52118         142.503221
#>   5:           76.81923           81.34900          89.894512
#>  ---                                                         
#> 125:         -142.55612          153.46214          83.465854
#> 126:          169.45247          125.76661         102.326258
#> 127:          -74.86515         -126.07900        -166.621579
#> 128:           46.28976           54.27060           1.254072
#> 129:          -56.93289         -170.52107         105.924844
#>      UVVIS_fft_phase_54 UVVIS_fft_phase_55 UVVIS_fft_phase_56
#>                   <num>              <num>              <num>
#>   1:          21.990235         -23.562288         -57.976297
#>   2:           6.040637          17.280337          36.209193
#>   3:         -53.830053         -86.571513         -93.489869
#>   4:         101.236276          39.080148           1.109029
#>   5:          50.802402          45.355371          51.760592
#>  ---                                                         
#> 125:         -36.182905         -59.278916         -83.400550
#> 126:         139.057318        -135.402599        -149.076358
#> 127:        -115.282255        -107.991158        -131.828491
#> 128:          13.457536         -38.869256         -84.373642
#> 129:          39.767995           3.357159           0.598408
#>      UVVIS_fft_phase_57 UVVIS_fft_phase_58 UVVIS_fft_phase_59
#>                   <num>              <num>              <num>
#>   1:        -135.017547          -82.11904          -89.81527
#>   2:           4.811745          -12.33403          -16.22858
#>   3:         -10.519104          -37.19529         -104.53285
#>   4:         -70.463110         -154.84046          154.08118
#>   5:          26.927057          -19.18503          -71.46789
#>  ---                                                         
#> 125:         -66.600881          -43.31322          -58.64287
#> 126:         154.215679         -175.10774         -138.51963
#> 127:        -131.674203         -144.05854         -172.78974
#> 128:        -153.456327          154.65567          112.01845
#> 129:         -10.557127          -25.59931          -28.91508
#>      UVVIS_fft_phase_60 UVVIS_fft_phase_61 UVVIS_fft_phase_62
#>                   <num>              <num>              <num>
#>   1:       -88.59390759        -97.3560516        -158.569384
#>   2:       -12.65851410         14.9985596          -9.586055
#>   3:         0.07521652         13.4570337          -6.239134
#>   4:       106.63010305         80.1442966         -90.873812
#>   5:        17.00705725          0.1325104         -34.399668
#>  ---                                                         
#> 125:       -81.54332256        -64.5340585          -7.077546
#> 126:      -174.18307879        151.9007087         118.645483
#> 127:       124.42308649         39.9661779        -103.119889
#> 128:        61.16543213         66.5170736          67.949005
#> 129:         4.49073500         17.3069230           9.712548
#>      UVVIS_fft_phase_63 UVVIS_fft_phase_64 UVVIS_fft_phase_65
#>                   <num>              <num>              <num>
#>   1:          166.14168         -168.77637        -152.416011
#>   2:         -138.90767           65.81918          12.386260
#>   3:          -49.81900           49.10798           8.386231
#>   4:         -137.83986          170.51523         104.206047
#>   5:          -76.08496         -138.18728          15.014400
#>  ---                                                         
#> 125:          -16.30264          -67.96918        -100.210407
#> 126:         -120.21570          -97.81169        -118.101096
#> 127:         -113.86558         -109.28662        -121.650637
#> 128:          -18.11523          -66.38954        -100.415491
#> 129:           29.62268           64.59808          24.667356
#>      UVVIS_fft_phase_66 UVVIS_fft_phase_67 UVVIS_fft_phase_68
#>                   <num>              <num>              <num>
#>   1:      -126.70413212         -177.78005      -1.384311e-11
#>   2:         7.48726635           10.69029      -1.348150e-13
#>   3:       -25.19347078          -37.75191      -1.126367e-13
#>   4:        -0.05453577          -21.42652      -8.633559e-14
#>   5:       -30.33804069          -39.24707      -1.780370e-12
#>  ---                                                         
#> 125:      -161.59285965          163.18044      -1.638134e-12
#> 126:      -135.06039923         -163.36831      -1.800000e+02
#> 127:      -147.40698030         -165.36687      -1.800000e+02
#> 128:      -161.98497954          163.10407      -4.801913e-13
#> 129:       -12.25026715          -25.05276      -1.759177e-13